+++ src/components/project/ProjectDetailsView/index.tsx
@@ -1,4 +1,4 @@
-import NegotiationHistoryTab from "@/components/project/components/Tab/NegotiationHistoryTabs";
+import NegotiationHistoryTab from "@/components/project/components/NegotiationHistoryTab/NegotiationHistoryTabs";
 import { Box } from "@mui/material";
 
 function ProjectDetailsView() {
